Refactor default module for Teacher model
